% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{getTestvector}
\alias{getTestvector}
\title{Create test vector}
\usage{
getTestvector(obj, eps = 1e-12)
}
\arguments{
\item{obj}{mboost object}

\item{eps}{numeric; value to stabilize inversion}
}
\value{
list of test vectors or matrices for each effect
}
\description{
Create test vector
}
