% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.iboost}
\alias{print.iboost}
\title{Generic methods for iboost objects}
\usage{
\method{print}{iboost}(obj)
}
\arguments{
\item{obj}{iboost object}
}
\description{
Generic methods which can be used for objects fitted with the \code{iboost} function
}
